add SFE MicroOLED display
Add a SparkFun Electronics MicroOLED display to the list of output devices. This includes an update to the macros.h file as well as a new menuIO file and an example for it's use.  Additional details, including the library, are available at https://github.com/sparkfun/SparkFun_Micro_OLED_Arduino_Library